A different take on sample db
This is a modification of Andrew's sampledb.sql that adds a few things mentioned in Issues and incorporates some other changes that might be useful. I'm adding it as a separate file, rather than a commit on Andrew's, because it's just an experiment for consideration. I have not yet included all the characters and states from Andrew's, nor the specimen.

Differences from Andrew's sampledb.sql:
-- Uses the APOC plugin to generate uuids
-- Adds Schema nodes and CITED_BY relationships between Reference and Character nodes
-- Keeps track author order
-- Relationship names changed to verbs rather than attributive nouns
-- Flips polarity on a numer of relationships
-- Changes SCHEMA relationships to CHARACTER_OF and STATE_OF
-- Adds Substate nodes and SUBSTATE_OF relationships