Update PBOT.cypher
Claire adds to Rebecca's work on schema MLA 2009 coding entries. I have added the code for venation and teeth. The MLA 2009 coding is now completed for Andrew and Doug's review. Smithsonian 1999 is NOT finished, it is still in progress. Perhaps we should wait for 2009 to be approved before finishing that code. Most of the code from 2009 should be able to be reused.

Becca's code starts on 422 and Claire's code starts on line 786.

NOTE: there are template lines at the bottom of this script that should be removed from the code since Claire is just using them to copy and paste in new lines of code as I work. 

Claire's questions for Andrew and Doug
I added a definition to a character--is that ok? For example, perimarginal veins
Can we do subscript and superscript, for example "lm" the m should be subscript for leaf margin length
How often do you have to commit to ensure not losing the progress you have made?
How can code handle when a character allows for the selection of multiple states? Ex. Epidemial tertiaries
Andrew, did I handle MLA Measurements ok? It's not technically a character in MLA, but this is data we would want to allow users to enter
Can we leave in the possibility for super characters? For example if someone consolidates current characters under a new heading. Also, would it be good to include 0.0.0.1.1.2.... to leave space for hierarchies expanding in both directions? Or I guess just update that part of the code as time comes????
Can we add images to the character state definitions? If so, botanists could capture the images into a folder and give them file names as appropriate then add the filenames into the code as appropriate.
How can we start to map out the skip logic for you? For example, if a person enters a particular trait it brings you to another linked trait. Also for example, allow folks to skip lower venation states if their specimen/morphotype only has up to 2nd order venation.

Claire's coding questions/comments for consideration by paleobotanists. 
Character Trait 6 Petiole Features. Should Petiole Base, Glands, Petiole-cross section, and Phyllodes be combined into Character 6 Petiole Features?
Should we add the MLA character number to Character State information for quick reference when doing data entry--especially for beginners who want to jump to 
that section of the manual?
See tooth shape code starting line 837 and make sure we like the way I handled it?
I left out advice from definitions to reduce length, for example "may be difficult to distinguish from...", "they are typically..."
33.3 and 33.4 are traits that appear to redefine earlier definitions of monopodial and sympodial--why not just use those words instead of redefining?
I added MLA Measured Traits as a character as well with quantitative entries--are we good with this?
Please help me understand intercostal tertiary vein variability, tertiary angle with what? see MLA
It appears that MLA uses multiple words for same meaning. Can we be consistent? For example always use absent instead of sometimes using lacking; polygons vs net?; 1st order vs 1° primary?